plugin Sigfox
========
- fix snr value.

+++ sources/plugins/Sigfox/specificHistorizer/SNR.cpp
@@ -5,10 +5,11 @@
 
 namespace specificHistorizers
 {
+   // TODO : To create Db unit
    DECLARE_CAPACITY(SNRCapacity, "SNR", shared::plugin::yPluginApi::CStandardUnits::NoUnits, shared::plugin::yPluginApi::EKeywordDataType::kNoData);
 
    CSNR::CSNR(const std::string& keywordName)
-      : CSingleHistorizableData<int>(keywordName,
+      : CSingleHistorizableData<double>(keywordName,
                                      SNRCapacity(),
                                      shared::plugin::yPluginApi::EKeywordAccessMode::kGet)
    {
